Clean up warnings

+++ src/graphics/glcontroller.cuh
@@ -32,7 +32,7 @@ namespace graphics
 
 		// Uniform matrices
 		glm::mat4 model = glm::scale(glm::mat4(1.0f), glm::vec3(0.5f, 0.5f, 0.5f));
-		glm::mat4 projection = glm::perspective(glm::radians<float>(45.0f), windowWidth / windowHeight, 0.1f, depth * 10);
+		glm::mat4 projection = glm::perspective(glm::radians<float>(45.0f), static_cast<float>(windowWidth) / windowHeight, 0.1f, depth * 10);
 
 		Model particleModel = Model("Models/Earth/low_poly_earth.fbx");
     Model veinModel = Model("Models/Cylinder/cylinder.obj");
